DNA_Sequencing_Using_BERT/dev_requirements.txt
scikit-learn
matplotlib
seaborn
pandas
Pillow